add msvc patch
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@2642 d0c010d8-bb82-477e-a075-112140230451

+++ src/CWpt.cpp
@@ -958,8 +958,8 @@ void CWpt::showBuddies(bool show)
 
             GPS_Math_Str_To_Deg(string, co.lon, co.lat, true);
 
-            co.lon *= DEG_TO_RAD;
-            co.lat *= DEG_TO_RAD;
+            co.lon *= (float) DEG_TO_RAD;
+            co.lat *= (float) DEG_TO_RAD;
 
             buddies << co;
         }
